pastebin_lite/requirements.txt
asgiref==3.11.0
dj-database-url==3.1.0
Django==6.0.1
djangorestframework==3.16.1
gunicorn==24.1.1
packaging==26.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
whitenoise==6.11.0
